requirements.txt
pandas
numpy
scikit-learn
tensorflow
streamlit
Pillow